refactor(codegen): split static-expression analysis out of patch_flag
Move the constant/static expression predicates (`is_constant_interpolation`,
`is_const_handler`, `is_static_bound_expression`, and their oxc-based literal
helpers `is_string_literal` / `is_static_object_or_array_literal` /
`is_static_oxc_expression`) out of patch_flag.rs into a new `patch_flag_expr`
module. They are pure, one-directional helpers; patch_flag.rs imports the three
entry points, and the oxc-parser imports move with them.

Pure code movement — no behavior change. patch_flag.rs shrinks 550 -> 417.
Verified across vize_atelier_dom, vize_atelier_sfc (528), vize_atelier_ssr (76),
and vize_atelier_core; clippy and rustfmt clean.
